middleware hanged on handlers; added DB interface and it's in memory realization for worker and vacancy entities

+++ clean-arch/internal/pkg/employer/employer.go
@@ -6,6 +6,7 @@ import "github.com/go-park-mail-ru/2024_2_VKatuny/clean-arch/internal/pkg/models
 // Now implemented as a in memory db.
 // Implementation locates in ./repository
 type Repository interface {
-	Create(*models.Employer) (uint32, error)
-	GetByID(id uint32) (*models.Employer, error)
+	//rename to Add
+	Create(*models.Employer) (uint64, error)
+	GetByID(id uint64) (*models.Employer, error)
 }
